Refactor spacecraft
